Set species flag on ck3/xk3->pk3
Make GetG3Species return ushort

Fixes regression caused by https://github.com/kwsch/PKHeX/commit/5942a741471e9b175fe07a9c2c90703f5dfe9d76
Copying SpeciesID3 instead of Species is fine as it skips the map/unmap step, but it also skips the setting of the HasSpecies flag. So we'll just set it in the individual ConvertToPK3 methods.

+++ PKHeX.Core/PKM/Util/Conversion/SpeciesConverter.cs
@@ -25,7 +25,7 @@ public static class SpeciesConverter
     /// </summary>
     /// <param name="species">National Dex ID</param>
     /// <returns>Generation 3 species ID.</returns>
-    public static int GetG3Species(int species) => (uint)species >= table3_Internal.Length ? 0 : table3_Internal[species];
+    public static ushort GetG3Species(int species) => (uint)species >= table3_Internal.Length ? (ushort)0 : table3_Internal[species];
 
     /// <summary>
     /// Converts Generation 3 species ID to National Dex ID.
